NEW Filter out more potentially sensitive vars (fixes #53)
This is technically an API change,
but if you really wanted to include those,
setFilterParams() can be used to remove those additional filters.
In 99% of cases, values for those keys weren't meant to be leaked
to third party systems such as Raygun.

As it turns out, this is a forward port from 1.x,
see https://github.com/silverstripe/silverstripe-raygun/pull/54